Backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.11.0
django-filter==21.1
django-import-export==2.8.0
django-storages==1.12.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.1.1
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
